pyproject.toml
[project]
name = "invoice-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai>=1.54.0",
    "pypdf>=5.1.0",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.1",
    "pymupdf>=1.25.0",
]
